Confused by "overwritten"
I honestly don't know if my changes are even correct, but I was quite confused by "overwritten" and thought setting the dev property didn't do much, since it will be "overwritten by the nuxt commands", and I just didn't understand what it was for. 

Also I didn't know what `production` and `development` mode actually meant and that this is a way to get `console.log()` back to figure out what's happening behind the curtains.

Please see the suggested changes more of a conversation starter than a perfect solution how to describe things. In the end it's just me who heavily misunderstood this property.